back-end/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pymysql 
pydantic
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]